Create reusbaleTableViewCell Fill up examples properties in Movie model; Create example VC for visible effect of cell;

+++ MovieMood/Models/Movie.swift
@@ -6,5 +6,5 @@ struct Movie {
     var movieCategory: String
     var movieDuration: String
     var movieRating: Float?
-    var movieRatingVoits: Float?
+    var movieRatingVoits: Int?
 }
